Added neural network infrastructure.
The network is snapshotted every 10000 minibatches from the Torch
trainer, and is sent back to the bot through the Python server to
be used with Neuroph. Keep in mind Neuroph and Torch don't really
play nice together, so a bit of data finagling is still needed on
the bot to get things to work properly.